Moves pdf_generation to services.
