+++ app/mailers/invoice_mailer.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 class InvoiceMailer < ApplicationMailer
-  include InvoicesHelper
-
   after_action -> { @invoice.sent! }
 
   def invoice
@@ -11,23 +9,7 @@ class InvoiceMailer < ApplicationMailer
     subject = params[:subject]
     @message = params[:message]
 
-    formatted_invoice = format_invoice(@invoice.invoice_line_items)
-
-    controller = ActionController::Base.new
-    html = controller.render_to_string(
-      template: "invoices/pdf",
-      layout: "layouts/pdf",
-      locals: {
-        invoice: @invoice,
-        company_logo:,
-        client: @invoice.client,
-        invoice_line_items: formatted_invoice[:invoice_line_items],
-        sub_total: formatted_invoice[:sub_total],
-        total: formatted_invoice[:total]
-      }
-      )
-
-    pdf = Grover.new(html).to_pdf
+    pdf = InvoicePayment::PdfGeneration.process(@invoice, company_logo)
     attachments["invoice_#{@invoice.invoice_number}.pdf"] = pdf
 
     mail(to: recipients, subject:, reply_to: "no-reply@miru.com")
